requirements.txt
Flask==2.0.1
Flask_SocketIO==5.2.0
numpy==1.21.2
opencv_python==4.5.3.56
pandas==1.3.2
Pillow==9.1.1
torch==1.9.0+cu111
